+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "start": "node ./bin/www",
     "build": "webpack",
-    "test": "NODE_ENV=test tape -r babel-register -r ./client/js/test/setup-browser-env.js ./client/js/test/",
+    "test": "NODE_ENV=test tape -r babel-register -r ./test/client/js/setup-browser-env.js ./test/client/js/",
     "cover": "nyc --reporter html --reporter cobertura --reporter json-summary npm test"
   },
   "dependencies": {
