+++ test/client/js/map.spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {shallow} from 'enzyme';
 import test from 'tape-catch';
 import sinon from 'sinon';
-import Map from '../map';
+import Map from '../../../client/js/map';
 
 test('Mounting the Map', assert => {
   document.body.innerHTML = `<button id="lgis-show-table">button</button>`;
