+++ test/client/js/smoke-test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import test from "tape-catch";
 
-import Map from "../map";
+import Map from "../../../client/js/map";
 
 test("Smoketests", assert => {
   const app = <Map />;
